--- src/main/java/io/zipcoder/casino/BlackJackHumanPlayer.java
@@ -1,21 +1,0 @@
-package io.zipcoder.casino;
-
-public class BlackJackHumanPlayer extends CardPlayer {
-
-
-    public BlackJackHumanPlayer(String name, Wallet gamblingWallet) {
-        super(name, gamblingWallet);
-    }
-
-    public Boolean doubleDown(){
-        // if true double wager
-        return null;
-    }
-
-    public Boolean split(){
-        //player has option to split only if both cards match
-
-        return null;
-    }
-
-}
