+++ src/main/java/io/zipcoder/casino/gamePlayers/BlackJackHumanPlayer.java
@@ -1,5 +1,24 @@
 package io.zipcoder.casino.gamePlayers;
 
-public class BlackJackHumanPlayer
-{
+import io.zipcoder.casino.CardPlayer;
+import io.zipcoder.casino.Wallet;
+
+public class BlackJackHumanPlayer extends CardPlayer {
+
+
+    public BlackJackHumanPlayer(String name, Wallet gamblingWallet) {
+        super(name, gamblingWallet);
+    }
+
+    public Boolean doubleDown(){
+        // if true double wager
+        return null;
+    }
+
+    public Boolean split(){
+        //player has option to split only if both cards match
+
+        return null;
+    }
+
 }
